fix(security): close authz gaps, panics, split-brain, and SQLite bind bugs
Backend hardening from a correctness + security bug hunt. Remote
`cargo check --workspace` + `cargo check -p machina-agent` pass; web
typecheck clean.

Authorization (daemon):
- automation.rs token/role create+list now admin-only; webhooks/
  notifications/alerts/schedules write-role gated (closes privilege
  escalation: any authenticated user could mint an admin API token).
- host_network.rs (bridges/port-forwards/firewall rules), vms.rs power
  ops, and k8s.rs destructive handlers now require can_write().
- consolehub end() verifies ownership before removing the session.

Panics / crashes:
- chunk_text (ai/mod.rs) and sniff_cloud_config_from_iso (agent) sliced
  at raw byte offsets → UTF-8 boundary panics (the agent one poisoned the
  libvirt mutex). Now char-boundary safe.
- overlay.rs /0 CIDR shift overflow; spec parse_memory_mib rejects 0/neg.
- k8s memory_mib overflow bounded.

HA correctness:
- ha.rs no longer recovers VMs off a host that wasn't confirmed fenced
  (split-brain); worker heartbeat clears the fenced flag so repeat
  failures re-fence.

SQLite migration fallout (controller):
- 4 swapped .bind() UPDATEs (console approval, spectator, replay, SIEM
  last_error) and 5 stringified-UUID-vs-BLOB queries fixed.
- operations_scheduler + fleet_snapshot_scheduler now leader-guarded.
- enqueue_task marks the row failed if the bus publish fails.

Agent security (contained; full gRPC auth still TODO):
- fence_host rejects shell metacharacters; restore/backup/clone disk
  paths confined to storage pools; provision_network name sanitized;
  serial console aborts the surviving task on close.

Other: SIEM exporters verify TLS by default (opt-out flag); openstack
image upload respects session cloud; balloon handler validates+routes;
console permission gate fails closed.

+++ core/src/network/overlay.rs
@@ -94,8 +94,9 @@ pub fn ip_from_cidr_offset(cidr: &str, offset: u32) -> Result<String, String> {
     if host_bits == 0 {
         return Err("CIDR has no host bits".into());
     }
-    let max_hosts = (1u32 << host_bits).saturating_sub(2);
-    if offset == 0 || offset > max_hosts {
+    // Widen to u64 so a /0 CIDR (host_bits == 32) doesn't overflow `1 << 32`.
+    let max_hosts = (1u64 << host_bits).saturating_sub(2);
+    if offset == 0 || u64::from(offset) > max_hosts {
         return Err(format!(
             "offset {offset} out of range for {cidr} (max {max_hosts})"
         ));
